Refactoring assembler into a general haplotype based caller.
